day3_context_memory/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

